add WIP identifyCoin wrapper
