+++ src/dart_interface.cpp
@@ -18,7 +18,7 @@
 using namespace spark;
 
 /*
- * FFI-friendly wrapper for spark:getAddress.
+ * FFI-friendly wrapper for spark::getAddress.
  */
 EXPORT_DART
 const char* getAddress(const char* keyDataHex, int index, int diversifier, int isTestNet) {
@@ -41,5 +41,33 @@ const char* getAddress(const char* keyDataHex, int index, int diversifier, int i
     } catch (const std::exception& e) {
         std::cerr << "Exception: " << e.what() << std::endl;
         return nullptr;
+    }
+}
+
+/*
+ * FFI-friendly wrapper for spark:identifyCoin.
+ *
+ * Uses the utility functions spark::Coin fromFFI(const CCoin& c_struct) to pass parameters to the
+ * C++ function spark::identifyCoin(const Coin& coin), then uses the utility function
+ * CIdentifiedCoinData toFFI(const spark::IdentifiedCoinData& cpp_struct) to convert the result back
+ * to a C struct.
+ *
+ * We also need the incoming view key or we need to derive it, so accept keyDataHex and index.
+ */
+EXPORT_DART
+CIdentifiedCoinData identifyCoin(const CCoin& c_struct, const char* keyDataHex, int index) {
+    try {
+        spark::Coin coin = fromFFI(c_struct);
+
+        // Derive the incoming view key from the key data and index.
+        spark::SpendKey spendKey = createSpendKeyFromData(keyDataHex, index);
+        spark::FullViewKey fullViewKey(spendKey);
+        spark::IncomingViewKey incomingViewKey(fullViewKey);
+
+        spark::IdentifiedCoinData identifiedCoinData = coin.identify(incomingViewKey);
+        return toFFI(identifiedCoinData);
+    } catch (const std::exception& e) {
+        std::cerr << "Exception: " << e.what() << std::endl;
+        return CIdentifiedCoinData();
     }
 }
